Moved pause menu view
